Loosen restriction on mailing list files
The first field is no more required to be the `email` field. The
`email` field can be anywhere - however, it nevertheless is
required.
Another change that was made was to the mailingList_bad.ml test
file. Originally, it was bad because the first field wasn't
email field - everything else was fine. Changed it to have
mismatch in number of columns.